+++ lib/sleek_circular_slider.dart
@@ -4,7 +4,7 @@ export 'src/circular_slider.dart';
 export 'src/settings/CircularSliderSettings.dart';
 export 'src/settings/CircularSliderFeatures.dart';
 export 'src/settings/CircularSliderColors.dart';
-export 'src/settings/CircularSliderValues.dart';
+export 'src/circular_slider_values.dart';
 export 'src/settings/CircularSliderShadow.dart';
 export 'src/settings/CircularSliderText.dart';
 export 'src/circular_slider_parameters.dart';
