+++ lib/src/circular_slider_values.dart
